DeliveryThree/Backend/requirements.txt
alembic==1.11.1
aniso8601==9.0.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
Flask==2.3.2
Flask-RESTful==0.3.9
#greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
#numpy==1.23.3
paho-mqtt==1.6.1
#pandas==1.5.0
pymongo==4.2.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.3